bot/requirements.txt
asyncio==3.4.3
python-dotenv==1.0.1
aiogram==3.8.0
transformers==4.42.3
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pytest==8.2.2
aio_pika==9.4.2
pillow==10.3.0
torch==2.3.1
